pyproject.toml
[project]
name = "url-shortner-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10,<3.12"
dependencies = [
    "cassandra-driver>=3.29.3",
    "fastapi>=0.121.0",
    "hashids>=1.3.1",
    "pydantic-settings>=2.11.0",
    "redis>=7.0.1",
    "uvicorn[standard]>=0.38.0",
]
